migrate getDateTypes to a static method mergeGraogQlDataTypes
