+++ packages/data-types/test/graphql-types-spec.ts
@@ -1,57 +1,6 @@
 
-import { getGraphQlTypes, DataType, DataTypeConfig } from '../src';
+describe('mergeGraphQLDataTypes', () => {
+    it('make a test here for graphql types', () => {
 
-describe('getGraphQlTypes', () => {
-    it('can build a single graphql type from multiple types', () => {
-        const typeConfig1: DataTypeConfig = {
-            version: 1,
-            fields: {
-                hello: { type: 'Text' },
-                location: { type: 'Geo' },
-                date: { type: 'Date' },
-                ip: { type: 'IP' },
-                someNum: { type: 'Long' }
-            },
-        };
-
-        const typeConfig2: DataTypeConfig = {
-            version: 1,
-            fields: {
-                hello: { type: 'Text' },
-                location: { type: 'Geo' },
-                otherLocation: { type: 'Geo' },
-                bool: { type: 'Boolean' }
-            },
-        };
-
-        const types = [
-            new DataType(typeConfig1, 'firstType'),
-            new DataType(typeConfig2, 'secondType'),
-        ];
-
-        const results = getGraphQlTypes(types);
-
-        const fields = [
-            'type firstType {',
-            'hello: String',
-            'location: Geo',
-            'date: DateTime',
-            'ip: String',
-            'someNum: Int',
-
-            'type secondType {',
-            'otherLocation: Geo',
-            'bool: Boolean',
-
-            'type Geo {',
-            'lat: String!',
-            'lon: String!'
-        ];
-
-        fields.forEach((str: string) => {
-            expect(results.match(str)).not.toBeNull();
-        });
-
-        expect(results.match(/type Geo \{/g)).toBeArrayOfSize(1);
     });
 });
